VMS-10817: Fixed 'Add Device' dialog layout and size constraints
VMS-10892, VMS-10893: Fixed appearance of server dropdown button in 'Add Device' dialog
VMS-11620: Fixed layout and validator hint style in 'Add Device' dialog
VMS-12529: Hide some of VA option for DW TVT VA camera. Another manifest bug fixed
VMS-12529: Hide some of VA option for DW TVT VA camera. Manifest bug fixed
VMS-12566: Device count in 'Add Devices' dialog table header updates properly
VMS-12906: allProperties/deviceDescriptors is changed in full-info after backup/restore operation
VMS-13194: Solve update tests hanging
VMS-13252: Camera/item hotkeys processing and indication
VMS-13271: Added HTTP CONNECT with authentication test
VMS-13271: Added number of comments and TODOs
VMS-13291: Fix several hanwha advanced paramites
VMS-13441: Removed unused class
VMS-13441: nx::network::http: Added convenience method validateAuthorization
VMS-13441: nx::network::http::AsyncClient: fixed: internal buffer was not cleared when opening new connection
VMS-13499: Clear all notifications only when logging out
VMS-13510: 'localhost' is resolved strongly to 127.0.0.1 by all applications that use QnStaticCommonModule
VMS-13571: Fixed onvif timestamp incorrect initialization
VMS-13576: Add 'disableMtDecoding' client option to force single thread video decoding
VMS-13601: Continued Dahua analytics events with action 'pulse' are processed only once
VMS-13617: Imaging settings are unavailable for IPC2122LR3-PF40M-D
VMS-13618: Added clearInputsTimeoutSec for K20IC
VMS-13625: Enable multicast automatically when RTP trasport = UDP Multicast
VMS-13661: Enable configurable session timeout feature
VMS-13662: Multiple soft triggers a appeared on item
VMS-13667: Prepend Ft prefix to the update test cases names
VMS-13667: Wait if the /ec2/updateInformation request didn't return an expected result at first attempt
VMS-13679: client cannot login server by url links
VMS-13680: Fixed crash if default rtp transport configured to 'auto'
VMS-13680: Fixed incorrect behavior when 'automatic' rtp transport used
VMS-13680: Reverted getTransport getter, fixed inacurate socket buffer size configuration
VMS-13681: Incorrect Japanese Translation (Branding violation)
VMS-13699: Add abort on cmd error
VMS-13699: Fix testcamera discovery search for all ifaces, including loopback
VMS-13699: Fix testcamera ignoring -I option
WEB-950: Web client doesn't work for all new web users
WEB-952: User lockout message is shown even after first try with incorrect password when no actual lockout happen
WEB-972: Realm is not after filled if create Cloud user from portal
WEB-975: Change text in webadmin for developer section

<==== Mercurial data ====>
Branch: vms

+++ vms/distribution/dmg/client/build_distribution.sh
@@ -41,7 +41,10 @@ patchDsStore()
 
 hardDetachDmg() # file.dmg
 {
-    lsof -t "$1" |xargs kill -9
+    if [ -f "$1" ]
+    then
+        lsof -t "$1" |xargs kill -9
+    fi
 }
 
 setDmgIcon()
